+++ aiohomematic/model/data_point.py
@@ -764,7 +764,7 @@ class BaseParameterDataPoint[
         "_ignore_on_initial_load",
         "_is_forced_sensor",
         "_is_un_ignored",
-        "_label",
+        "_translation",
         "_max",
         "_min",
         "_multiplier",
@@ -786,7 +786,7 @@ class BaseParameterDataPoint[
         "_translation_key",
         "_type",
         "_unit",
-        "_value_labels",
+        "_value_translations",
         "_values",
         "_visible",
     )
@@ -825,7 +825,7 @@ class BaseParameterDataPoint[
             parameter=self._parameter,
             custom_only=True,
         )
-        self._label: Final[str | None] = ccu_translations.get_parameter_label(
+        self._translation: Final[str | None] = ccu_translations.get_parameter_translation(
             parameter=self._parameter,
             channel_type=channel.type_name,
             locale=channel.device.config_provider.config.locale,
@@ -840,9 +840,9 @@ class BaseParameterDataPoint[
         self._state_uncertain: bool = True
         self._is_forced_sensor: bool = False
         self._assign_parameter_data(parameter_data=parameter_data)
-        self._value_labels: Final[dict[str, str | None] | None] = (
+        self._value_translations: Final[dict[str, str | None] | None] = (
             {
-                v: ccu_translations.get_parameter_value_label(
+                v: ccu_translations.get_parameter_value_translation(
                     parameter=self._parameter,
                     value=v,
                     channel_type=channel.type_name,
@@ -884,7 +884,6 @@ class BaseParameterDataPoint[
     ignore_on_initial_load: Final = DelegatedProperty[bool](path="_ignore_on_initial_load")
     is_forced_sensor: Final = DelegatedProperty[bool](path="_is_forced_sensor")
     is_un_ignored: Final = DelegatedProperty[bool](path="_is_un_ignored")
-    label: Final = DelegatedProperty[str | None](path="_label", kind=Kind.INFO)
     last_non_default_value: Final = DelegatedProperty[ParameterT | None](path="_last_non_default_value")
     max: Final = DelegatedProperty[ParameterT](path="_max", kind=Kind.CONFIG)
     min: Final = DelegatedProperty[ParameterT](path="_min", kind=Kind.CONFIG)
@@ -896,9 +895,12 @@ class BaseParameterDataPoint[
     status: Final = DelegatedProperty[ParameterStatus | None](path="_status_value")
     status_dpk: Final = DelegatedProperty[DataPointKey | None](path="_status_dpk")
     status_parameter: Final = DelegatedProperty[str | None](path="_status_parameter")
+    translation: Final = DelegatedProperty[str | None](path="_translation", kind=Kind.INFO)
     translation_key: Final = DelegatedProperty[str](path="_translation_key")
     unit: Final = DelegatedProperty[str | None](path="_unit", kind=Kind.CONFIG)
-    value_labels: Final = DelegatedProperty[dict[str, str | None] | None](path="_value_labels", kind=Kind.CONFIG)
+    value_translations: Final = DelegatedProperty[dict[str, str | None] | None](
+        path="_value_translations", kind=Kind.CONFIG
+    )
     values: Final = DelegatedProperty[tuple[str, ...] | None](path="_values", kind=Kind.CONFIG)
     visible: Final = DelegatedProperty[bool](path="_visible")